Fix option type

+++ src/subcommand/giraffe_main.cpp
@@ -748,7 +748,7 @@ int main_giraffe(int argc, char** argv) {
         .add_entry<double>("fragment-set-score-threshold", std::numeric_limits<double>::max())
         .add_entry<int>("min-chaining-problems", 1)
         .add_entry<int>("max-chaining-problems", std::numeric_limits<int>::max())
-        .add_entry<int>("max-lookback-bases", 24000)
+        .add_entry<size_t>("max-lookback-bases", 24000)
         .add_entry<int>("min-chains", 4)
         .add_entry<size_t>("max-chains-per-tree", 5)
         .add_entry<size_t>("max-alignments", 5);
